mcp/requirements.txt
mcp
psycopg2-binary
python-dotenv
httpx